Cast to size_t

+++ src/ctfeexpr.c
@@ -470,7 +470,7 @@ ArrayLiteralExp *createBlockDuplicatedArrayLiteral(Loc loc, Type *type,
     {
         // If it is a multidimensional array literal, do it recursively
         elem = createBlockDuplicatedArrayLiteral(loc, type->nextOf(), elem,
-            ((TypeSArray *)type->nextOf())->dim->toInteger());
+            (size_t)((TypeSArray *)type->nextOf())->dim->toInteger());
     }
     bool mustCopy = needToCopyLiteral(elem);
     for (size_t i = 0; i < dim; i++)
